Fix unary minus warnings
No longer disable wd4146, fix warnings, and disable this warning when
including LLVM headers in LLVMReactor.cpp, as it trips for specific
headers like MathExtras.h, StringExtras.h.

Fixes ANGLE roll.

Bug: angleproject:4034
Reviewed-on: https://swiftshader-review.googlesource.com/c/SwiftShader/+/37488
Kokoro-Presubmit: kokoro <noreply+kokoro@google.com>
Tested-by: Antonio Maiorano <amaiorano@google.com>

+++ src/WSI/Win32SurfaceKHR.cpp
@@ -131,7 +131,7 @@ void Win32SurfaceKHR::lazyCreateFrameBuffer()
 	bitmapInfo.bmiHeader.biSize = sizeof(BITMAPINFO);
 	bitmapInfo.bmiHeader.biBitCount = 32;
 	bitmapInfo.bmiHeader.biPlanes = 1;
-	bitmapInfo.bmiHeader.biHeight = -windowExtent.height;
+	bitmapInfo.bmiHeader.biHeight = -static_cast<LONG>(windowExtent.height); // Negative for top-down DIB, origin in upper-left corner
 	bitmapInfo.bmiHeader.biWidth = windowExtent.width;
 	bitmapInfo.bmiHeader.biCompression = BI_RGB;
